requirements.txt
google-api-python-client
oauth2client
transformers>=4.30.0
torch>=1.12.0
sentencepiece
scikit-learn
nltk
requests
python-dotenv
